Reworking the way the broadcast queue works for async behaviour
Instead of spinning in a loop, it'll use the new async WaitHandle extension
to wait for the queue to actually receive a message.

Also refactored the Android MainActivity so that the websocket abstraction
is in a separate Wrapper class which provides an OnMEssage event to
indicate that stuff is available to be read.

Note: didn't pass the actual messages back in OnMEssage in case the incoming
message events swamp the processor and messages get missed.
...but we'll also need an enumeration or something for that.