Binding the Status Model
So the only way I could get this to work was to add an explicit property to the ViewModel.
Usual issues though - do I really have to add the whole OnPropertyChanged() to every property?
Do I have to duplicate the HistoryEntry props in the view model?
Opted instead to store the whole HistoryEntry in the viewmodel, and use a generic "PRopertyNotifier" that takes as its type args
the type of object that we're hosting (in this case HistoryEntry) and the type of object that is wrapping it (in this case the StatusViewModel).
It finds all the common properties and calls notify on the target object.
To make this work, a new interface is needed - IPropertyChangedNotifer - which declares the OnPropertyChanged() method.
This is implemented by BaseView - which our JournalEntryViewModel extends.
When the LastEntry (the historyentry instance) is updated (i.e. when we replace the instance), we call the PropertyNotifier to signal that all the properties therein have changed.
We simply need to expose the data we want to view as readonly properties on the viewmodel and map them to the LastEntry.
(allowing for Null of course).